test: add private key input to e2e suite + keep all junit result files (#3747)

+++ test/e2e/cluster.sh
@@ -8,6 +8,9 @@ GOPATH="/go"
 WORK_DIR="/aks-engine"
 MASTER_VM_UPGRADE_SKU="${MASTER_VM_UPGRADE_SKU:-Standard_D4_v3}"
 AZURE_ENV="${AZURE_ENV:-AzurePublicCloud}"
+IDENTITY_SYSTEM="${IDENTITY_SYSTEM:-azure_ad}"
+GINKGO_FAIL_FAST="${GINKGO_FAIL_FAST:-false}"
+TEST_PVC="${TEST_PVC:-false}"
 mkdir -p _output || exit 1
 
 # Assumes we're running from the git root of aks-engine
@@ -35,6 +38,23 @@ if [ -n "$ADD_NODE_POOL_INPUT" ]; then
 ${ADD_NODE_POOL_INPUT}
 END
 fi
+
+if [ -n "$PRIVATE_SSH_KEY_FILE" ]; then
+  PRIVATE_SSH_KEY_FILE=$(realpath --relative-to=$(pwd) ${PRIVATE_SSH_KEY_FILE})
+fi
+
+function tryExit {
+  if [ "${AZURE_ENV}" != "AzureStackCloud" ]; then
+    exit 1
+  fi
+}
+
+function renameResultsFile {
+  JUNIT_PATH=$(pwd)/test/e2e/kubernetes/junit.xml
+  if [ "${AZURE_ENV}" == "AzureStackCloud" ] && [ -f ${JUNIT_PATH} ]; then
+    mv ${JUNIT_PATH} $(pwd)/test/e2e/kubernetes/${1}-junit.xml
+  fi
+}
 
 echo "Running E2E tests against a cluster built with the following API model:"
 cat ${TMP_DIR}/apimodel-input.json
@@ -84,6 +104,7 @@ docker run --rm \
 -e ORCHESTRATOR=kubernetes \
 -e ORCHESTRATOR_RELEASE="${ORCHESTRATOR_RELEASE}" \
 -e CREATE_VNET="${CREATE_VNET}" \
+-e PRIVATE_SSH_KEY_FILE="${PRIVATE_SSH_KEY_FILE}" \
 -e TIMEOUT="${E2E_TEST_TIMEOUT}" \
 -e LB_TIMEOUT="${LB_TEST_TIMEOUT}" \
 -e KUBERNETES_IMAGE_BASE=$KUBERNETES_IMAGE_BASE \
@@ -108,8 +129,9 @@ docker run --rm \
 -e LOG_ANALYTICS_WORKSPACE_KEY="${LOG_ANALYTICS_WORKSPACE_KEY}" \
 -e CUSTOM_HYPERKUBE_IMAGE="${CUSTOM_HYPERKUBE_IMAGE}" \
 -e IS_JENKINS="${IS_JENKINS}" \
+-e TEST_PVC="${TEST_PVC}" \
 -e SKIP_TEST="${SKIP_TESTS}" \
--e GINKGO_FAIL_FAST=true \
+-e GINKGO_FAIL_FAST="${GINKGO_FAIL_FAST}" \
 -e GINKGO_FOCUS="${GINKGO_FOCUS}" \
 -e GINKGO_SKIP="${GINKGO_SKIP}" \
 -e API_PROFILE="${API_PROFILE}" \
@@ -129,7 +151,7 @@ docker run --rm \
 -e GRAPH_ENDPOINT="${GRAPH_ENDPOINT}" \
 -e SERVICE_MANAGEMENT_VM_DNS_SUFFIX="${SERVICE_MANAGEMENT_VM_DNS_SUFFIX}" \
 -e RESOURCE_MANAGER_VM_DNS_SUFFIX="${RESOURCE_MANAGER_VM_DNS_SUFFIX}" \
-"${DEV_IMAGE}" make test-kubernetes || exit 1
+"${DEV_IMAGE}" make test-kubernetes || tryExit && renameResultsFile "deploy"
 
 if [ "${UPGRADE_CLUSTER}" = "true" ] || [ "${SCALE_CLUSTER}" = "true" ] || [ -n "$ADD_NODE_POOL_INPUT" ] || [ "${GET_CLUSTER_LOGS}" = "true" ]; then
   # shellcheck disable=SC2012
@@ -148,6 +170,7 @@ if [ "${UPGRADE_CLUSTER}" = "true" ] || [ "${SCALE_CLUSTER}" = "true" ] || [ -n 
   fi
   
   if [ "${GET_CLUSTER_LOGS}" = "true" ]; then
+      PRIVATE_SSH_KEY_FILE="${PRIVATE_SSH_KEY_FILE:-_output/${RESOURCE_GROUP}-ssh}"
       docker run --rm \
       -v $(pwd):${WORK_DIR} \
       -w ${WORK_DIR} \
@@ -158,7 +181,7 @@ if [ "${UPGRADE_CLUSTER}" = "true" ] || [ "${SCALE_CLUSTER}" = "true" ] || [ -n 
       --api-model _output/$RESOURCE_GROUP/apimodel.json \
       --location $REGION \
       --ssh-host $API_SERVER \
-      --linux-ssh-private-key _output/$RESOURCE_GROUP-ssh \
+      --linux-ssh-private-key $PRIVATE_SSH_KEY_FILE \
       --linux-script ./scripts/collect-logs.sh
       # TODO remove --linux-script once collect-logs.sh is part of the VHD
   fi
@@ -172,18 +195,21 @@ if [ "${UPGRADE_CLUSTER}" = "true" ] || [ "${SCALE_CLUSTER}" = "true" ] || [ -n 
       done
     done
   fi
-  git reset --hard
-  git remote rm $UPGRADE_FORK
-  git remote add $UPGRADE_FORK https://github.com/$UPGRADE_FORK/aks-engine.git
-  git fetch --prune $UPGRADE_FORK
-  git branch -D $UPGRADE_FORK/$UPGRADE_BRANCH
-  git checkout -b $UPGRADE_FORK/$UPGRADE_BRANCH --track $UPGRADE_FORK/$UPGRADE_BRANCH
-  git pull
-  git log -1
-  docker run --rm \
-    -v $(pwd):${WORK_DIR} \
-    -w ${WORK_DIR} \
-    "${DEV_IMAGE}" make build-binary > /dev/null 2>&1 || exit 1
+  
+  if [ "${UPGRADE_CLUSTER}" = "true" ]; then
+    git reset --hard
+    git remote rm $UPGRADE_FORK
+    git remote add $UPGRADE_FORK https://github.com/$UPGRADE_FORK/aks-engine.git
+    git fetch --prune $UPGRADE_FORK
+    git branch -D $UPGRADE_FORK/$UPGRADE_BRANCH
+    git checkout -b $UPGRADE_FORK/$UPGRADE_BRANCH --track $UPGRADE_FORK/$UPGRADE_BRANCH
+    git pull
+    git log -1
+    docker run --rm \
+      -v $(pwd):${WORK_DIR} \
+      -w ${WORK_DIR} \
+      "${DEV_IMAGE}" make build-binary > /dev/null 2>&1 || exit 1
+  fi
 else
   exit 0
 fi
@@ -232,9 +258,10 @@ if [ -n "$ADD_NODE_POOL_INPUT" ]; then
     -e REGIONS=$REGION \
     -e IS_JENKINS=${IS_JENKINS} \
     -e SKIP_LOGS_COLLECTION=true \
-    -e GINKGO_FAIL_FAST=true \
+    -e GINKGO_FAIL_FAST="${GINKGO_FAIL_FAST}" \
     -e GINKGO_SKIP="${SKIP_AFTER_SCALE_DOWN}" \
     -e GINKGO_FOCUS="${GINKGO_FOCUS}" \
+    -e TEST_PVC="${TEST_PVC}" \
     -e SKIP_TEST=${SKIP_TESTS_AFTER_ADD_POOL} \
     -e ADD_NODE_POOL_INPUT=${ADD_NODE_POOL_INPUT} \
     -e API_PROFILE="${API_PROFILE}" \
@@ -254,7 +281,7 @@ if [ -n "$ADD_NODE_POOL_INPUT" ]; then
     -e GRAPH_ENDPOINT="${GRAPH_ENDPOINT}" \
     -e SERVICE_MANAGEMENT_VM_DNS_SUFFIX="${SERVICE_MANAGEMENT_VM_DNS_SUFFIX}" \
     -e RESOURCE_MANAGER_VM_DNS_SUFFIX="${RESOURCE_MANAGER_VM_DNS_SUFFIX}" \
-    ${DEV_IMAGE} make test-kubernetes || exit 1
+    ${DEV_IMAGE} make test-kubernetes || tryExit && renameResultsFile "add-node-pool"
 fi
 
 if [ "${SCALE_CLUSTER}" = "true" ]; then
@@ -300,9 +327,10 @@ if [ "${SCALE_CLUSTER}" = "true" ]; then
     -e REGIONS=$REGION \
     -e IS_JENKINS=${IS_JENKINS} \
     -e SKIP_LOGS_COLLECTION=true \
-    -e GINKGO_FAIL_FAST=true \
+    -e GINKGO_FAIL_FAST="${GINKGO_FAIL_FAST}" \
     -e GINKGO_SKIP="${SKIP_AFTER_SCALE_DOWN}" \
     -e GINKGO_FOCUS="${GINKGO_FOCUS}" \
+    -e TEST_PVC="${TEST_PVC}" \
     -e SKIP_TEST=${SKIP_TESTS_AFTER_SCALE_DOWN} \
     -e ADD_NODE_POOL_INPUT=${ADD_NODE_POOL_INPUT} \
     -e API_PROFILE="${API_PROFILE}" \
@@ -322,7 +350,7 @@ if [ "${SCALE_CLUSTER}" = "true" ]; then
     -e GRAPH_ENDPOINT="${GRAPH_ENDPOINT}" \
     -e SERVICE_MANAGEMENT_VM_DNS_SUFFIX="${SERVICE_MANAGEMENT_VM_DNS_SUFFIX}" \
     -e RESOURCE_MANAGER_VM_DNS_SUFFIX="${RESOURCE_MANAGER_VM_DNS_SUFFIX}" \
-    ${DEV_IMAGE} make test-kubernetes || exit 1
+    ${DEV_IMAGE} make test-kubernetes || tryExit && renameResultsFile "scale-down"
 fi
 
 if [ "${UPGRADE_CLUSTER}" = "true" ]; then
@@ -380,9 +408,10 @@ if [ "${UPGRADE_CLUSTER}" = "true" ]; then
       -e REGIONS=$REGION \
       -e IS_JENKINS=${IS_JENKINS} \
       -e SKIP_LOGS_COLLECTION=${SKIP_LOGS_COLLECTION} \
-      -e GINKGO_FAIL_FAST=true \
+      -e GINKGO_FAIL_FAST="${GINKGO_FAIL_FAST}" \
       -e GINKGO_SKIP="${SKIP_AFTER_UPGRADE}" \
       -e GINKGO_FOCUS="${GINKGO_FOCUS}" \
+      -e TEST_PVC="${TEST_PVC}" \
       -e SKIP_TEST=${SKIP_TESTS_AFTER_UPGRADE} \
       -e ADD_NODE_POOL_INPUT=${ADD_NODE_POOL_INPUT} \
       -e API_PROFILE="${API_PROFILE}" \
@@ -402,7 +431,7 @@ if [ "${UPGRADE_CLUSTER}" = "true" ]; then
       -e GRAPH_ENDPOINT="${GRAPH_ENDPOINT}" \
       -e SERVICE_MANAGEMENT_VM_DNS_SUFFIX="${SERVICE_MANAGEMENT_VM_DNS_SUFFIX}" \
       -e RESOURCE_MANAGER_VM_DNS_SUFFIX="${RESOURCE_MANAGER_VM_DNS_SUFFIX}" \
-      ${DEV_IMAGE} make test-kubernetes || exit 1
+      ${DEV_IMAGE} make test-kubernetes || tryExit && renameResultsFile "upgrade"
   done
 fi
 
@@ -449,9 +478,10 @@ if [ "${SCALE_CLUSTER}" = "true" ]; then
     -e REGIONS=$REGION \
     -e IS_JENKINS=${IS_JENKINS} \
     -e SKIP_LOGS_COLLECTION=${SKIP_LOGS_COLLECTION} \
-    -e GINKGO_FAIL_FAST=true \
+    -e GINKGO_FAIL_FAST="${GINKGO_FAIL_FAST}" \
     -e GINKGO_SKIP="${SKIP_AFTER_SCALE_UP}" \
     -e GINKGO_FOCUS="${GINKGO_FOCUS}" \
+    -e TEST_PVC="${TEST_PVC}" \
     -e SKIP_TEST=${SKIP_TESTS_AFTER_SCALE_UP} \
     -e ADD_NODE_POOL_INPUT=${ADD_NODE_POOL_INPUT} \
     -e API_PROFILE="${API_PROFILE}" \
@@ -471,5 +501,5 @@ if [ "${SCALE_CLUSTER}" = "true" ]; then
     -e GRAPH_ENDPOINT="${GRAPH_ENDPOINT}" \
     -e SERVICE_MANAGEMENT_VM_DNS_SUFFIX="${SERVICE_MANAGEMENT_VM_DNS_SUFFIX}" \
     -e RESOURCE_MANAGER_VM_DNS_SUFFIX="${RESOURCE_MANAGER_VM_DNS_SUFFIX}" \
-    ${DEV_IMAGE} make test-kubernetes || exit 1
+    ${DEV_IMAGE} make test-kubernetes || tryExit && renameResultsFile "scale-up"
 fi
